+++ packages/fast-tooling-react/src/form/form/soft-remove.tsx
@@ -48,8 +48,8 @@ class SoftRemove extends React.Component<SoftRemoveProps, {}> {
         if (this.props.checked) {
             return (
                 <svg
-                    width="16"
-                    height="16"
+                    width="12"
+                    height="14"
                     viewBox="0 0 10 16"
                     xmlns="http://www.w3.org/2000/svg"
                 >
@@ -69,8 +69,8 @@ class SoftRemove extends React.Component<SoftRemoveProps, {}> {
 
         return (
             <svg
-                width="16"
-                height="16"
+                width="12"
+                height="14"
                 viewBox="0 0 10 16"
                 xmlns="http://www.w3.org/2000/svg"
             >
